Add a shell script to update SRCREV directives in bb files
To ease the pain of updateing the SRCREV lines in a lot of files for each Qt
update, here is a little shell script that does it for as. It also sets the
PV and QT_MODULE_BRANCH variables in recipes-qt/qt5/qt5-git.inc.

It is supposed to be called from the top-level directory of meta-qt5 with
the new version number as only argument. Like this, for instance:

	lib/update-qt-versions.sh 5.11.1

Signed-off-by: Daniel Mack <daniel@zonque.org>